Polish: extract shared Google auth/status, update README (#23)
